# Api-FarmaControl/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FarmaControl API",
    "version": "1.0.0",
    "description": "API para la gestión de farmacia: usuarios, productos, ventas, compras, proveedores y detalles."
  },
  "servers": [
    { "url": "http://localhost:3000/api" }
  ],
...
  "paths": {
    "/usuarios": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Obtener todos los usuarios",
        "responses": { "200": { "description": "Lista de usuarios" } }
      },
      "post": {
        "tags": ["Usuarios"],
        "summary": "Crear un usuario",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Usuario" } } }
        },
        "responses": { "201": { "description": "Usuario creado" } }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Obtener usuario por ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Usuario encontrado" }, "404": { "description": "Usuario no encontrado" } }
      },
      "put": {
        "tags": ["Usuarios"],
        "summary": "Actualizar usuario",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Usuario" } } }
        },
        "responses": { "200": { "description": "Usuario actualizado" } }
      },
      "delete": {
        "tags": ["Usuarios"],
        "summary": "Eliminar usuario",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Usuario eliminado" } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Usuario": {
        "type": "object",
        "properties": {
          "usuario": { "type": "string" },
          "contrasenia": { "type": "string" },
          "nombre_completo": { "type": "string" },
          "email": { "type": "string" },
          "rol": { "type": "string" }
        }
      },
...
    }
  }
}